+++ src/ResourceManager/AnalysisServices/Commands.AnalysisServices/Models/AzureAnalysisServicesServer.cs
@@ -15,7 +15,7 @@
 using System;
 using Microsoft.Azure.Management.Analysis.Models;
 using System.Collections.Generic;
-using Newtonsoft.Json;
+using Microsoft.Azure.Management.Internal.Resources.Utilities.Models;
 
 namespace Microsoft.Azure.Commands.AnalysisServices.Models
 {
@@ -38,6 +38,8 @@ namespace Microsoft.Azure.Commands.AnalysisServices.Models
         public string ServerFullName { get; private set; }
 
         public string BackupBlobContainerUri { get; set; }
+
+        public string ResourceGroupName { get; set; }
 
         public ServerSku Sku { get; set; }
 
@@ -79,6 +81,8 @@ namespace Microsoft.Azure.Commands.AnalysisServices.Models
                 config = new PsAzureAnalysisServicesFirewallConfig(enablePowerBIService, rules);
             }
 
+            ResourceIdentifier resourceIdentifier = new ResourceIdentifier(server.Id);
+
             return new AzureAnalysisServicesServer()
             {
                 AsAdministrators = server.AsAdministrators == null
@@ -94,6 +98,7 @@ namespace Microsoft.Azure.Commands.AnalysisServices.Models
                 Sku = server.Sku != null ? ServerSku.FromResourceSku(server.Sku) : new ServerSku(),
                 Tag = server.Tags != null ? new Dictionary<string, string>(server.Tags) : new Dictionary<string, string>(),
                 BackupBlobContainerUri = server.BackupBlobContainerUri == null ? String.Empty : server.BackupBlobContainerUri,
+                ResourceGroupName = resourceIdentifier.ResourceGroupName,
                 DefaultConnectionMode = server.QuerypoolConnectionMode.ToString(),
                 FirewallConfig = config,
                 GatewayInfo = server.GatewayDetails != null ? ServerGateway.FromResourceGateway(server.GatewayDetails) : null