+++ src/ResourceManager/Compute/Commands.Compute/Extension/NetcoreExtension.cs
@@ -58,7 +58,6 @@ namespace Microsoft.Azure.Commands.Compute
 #else
 using Microsoft.Azure.Commands.Common.Authentication;
 using Microsoft.Azure.Commands.Common.Authentication.Abstractions;
-using Microsoft.Azure.Management.Network;
 using Microsoft.Azure.Management.Storage.Models;
 using Newtonsoft.Json;
 using System;
@@ -68,6 +67,7 @@ using System.Collections.Generic;
 using System.Text;
 using System.Threading;
 using System.Threading.Tasks;
+using Microsoft.Azure.Management.Internal.Network.Version2017_10_01;
 
 namespace Microsoft.Azure.Commands.Compute
 {