+++ src/ResourceManager/Insights/Commands.Insights/ActivityLogAlert/NewAzureRmActionGroupCommand.cs
@@ -13,7 +13,7 @@
 // ----------------------------------------------------------------------------------
 
 using Microsoft.Azure.Commands.ResourceManager.Common;
-using Microsoft.Azure.Management.Monitor.Management.Models;
+using Microsoft.Azure.Management.Monitor.Models;
 using System.Management.Automation;
 using System.Collections.Generic;
 
@@ -22,7 +22,7 @@ namespace Microsoft.Azure.Commands.Insights.ActivityLogAlert
     /// <summary>
     /// Create an Activity Log Alert Action Group
     /// </summary>
-    [Cmdlet(VerbsCommon.New, "AzureRmActionGroup"), OutputType(typeof(ActivityLogAlertActionGroup))]
+    [Cmdlet("New", ResourceManager.Common.AzureRMConstants.AzureRMPrefix + "ActionGroup"), OutputType(typeof(Management.Monitor.Management.Models.ActivityLogAlertActionGroup))]
     public class NewAzureRmActionGroupCommand : AzureRMCmdlet
     {
         #region Cmdlet parameters
@@ -49,9 +49,11 @@ namespace Microsoft.Azure.Commands.Insights.ActivityLogAlert
         public override void ExecuteCmdlet()
         {
             WriteObject(
-                new ActivityLogAlertActionGroup(
-                    actionGroupId: this.ActionGroupId,
-                    webhookProperties: this.WebhookProperty));
+                new Management.Monitor.Management.Models.ActivityLogAlertActionGroup
+                {
+                    ActionGroupId = this.ActionGroupId,
+                    WebhookProperties = this.WebhookProperty
+                });
         }
     }
 }