+++ src/Storage/Commands.Storage/File/Cmdlet/SetAzureStorageShareQuota.cs
@@ -20,7 +20,7 @@ using System.Security.Permissions;
 
 namespace Microsoft.WindowsAzure.Commands.Storage.File.Cmdlet
 {
-    [Cmdlet(VerbsCommon.Set, StorageNouns.ShareQuota), OutputType(typeof(FileShareProperties))]
+    [Cmdlet("Set", Azure.Commands.ResourceManager.Common.AzureRMConstants.AzurePrefix + "StorageShareQuota"), OutputType(typeof(FileShareProperties))]
     public class SetAzureStorageShareQuota : AzureStorageFileCmdletBase
     {
         [Alias("N", "Name")]
@@ -43,7 +43,6 @@ namespace Microsoft.WindowsAzure.Commands.Storage.File.Cmdlet
 
         [Parameter(Position = 1, Mandatory = true,
             HelpMessage = "Share Quota")]
-        [ValidateRange(1, 5120)]
         public int Quota { get; set; }
 
         /// <summary>
@@ -80,4 +79,3 @@ namespace Microsoft.WindowsAzure.Commands.Storage.File.Cmdlet
         }
     }
 }
-