+++ src/ResourceManager/SignalR/Commands.SignalR/ChangeLog.md
@@ -19,6 +19,12 @@
 -->
 ## Current Release
 
+## Version 0.1.4
+* Updated to the latest version of the Azure ClientRuntime.
+
+## Version 0.1.3
+* Updated help files to include full parameter types and correct input/output types.
+
 ## Version 0.1.2
 * `New-AzureRmSignalR` uses a new version of `Strategy` library (4.0).